Module lazy_prime_sieve::sieve

Module providing sieves for generating primes.

Structs

Entry

Table entry for genuine sieve of eratosthenes.

GenuineSieve

Genuine Sieve of eratosthenes implementation based on the paper: The Genuine Sieve of Eratosthenes

TrialDivisionSieve

The modulus based memoized approach of generating primes that we all know and love.

UnfaithfulSieve

Non-Recursive Rust implementation of the classic Haskell recursive method for generating primes.

Enums

IterMultiple

Creates an Iterator of integer multiples from the given iterator.

Type Definitions

Table

Table for maintaining composites in the genuine prime sieve.